// src/interFace/interFace.ts
export interface MenuType {
  id: number;
  hasDropdown?: boolean;
  megaMenu?: boolean;
  lastDropdown?: boolean;
  active?: boolean;
  children?: boolean;
  title: string;
  pluseIncon?: boolean;
  link: string;
  previewImg?: boolean;
  pageLayout?: boolean;
  submenus?: any[];
  pages?: boolean;
}